+++ factories/demo_view_abstract_factory.py
@@ -1,1 +1,0 @@
-# [TODO] Design for similar widgets can be abstracted into a base class